make connect a single module
